updated nodes definitions
