+++ scripts/launch.sh
@@ -1,8 +1,8 @@
 #  Copyright (c) 2023. Tudor Oancea EPFL Racing Team Driverless
 if [ -z "$BRAINS_LAUNCH_FILE" ]; then
   if [ -z "$1" ]; then
-    echo "BRAINS_LAUNCH_FILE is empty, using launch_default.py"
-    BRAINS_LAUNCH_FILE="launch_default.py"
+    echo "BRAINS_LAUNCH_FILE is empty, using default_launch.py"
+    BRAINS_LAUNCH_FILE="default_launch.py"
   else
     BRAINS_LAUNCH_FILE=$1
   fi
@@ -16,5 +16,5 @@ if [ -z "$BRAINS_CONFIG_FILE" ]; then
   fi
 fi
 
-source install/setup.zsh
-ros2 pylaunch pylaunch "$BRAINS_LAUNCH_FILE"
+source install/setup.sh
+ros2 launch brains_launch "$BRAINS_LAUNCH_FILE"
