+++ src/brains_python/setup.py
@@ -22,9 +22,8 @@ setup(
     tests_require=["pytest"],
     entry_points={
         "console_scripts": [
-            "fsds_car_sensors = brains_python.nodes.fsds_car_sensors:main",
+            "fsds_velocity_estimation = brains_python.nodes.fsds_velocity_estimation:main",
             "fsds_car_controller = brains_python.nodes.fsds_car_controller:main",
-            "velocity_estimation = brains_python.nodes.velocity_estimation:main",
             "ekf_slam = brains_python.nodes.ekf_slam:main",
             "control_known_track = brains_python.nodes.control_known_track:main",
             "point_cloud_file = brains_python.nodes.point_cloud_file:main",
